Extract FunctionConfiguration header so that only dev server needs to pass it in
Cloud dashboard has its own header over the whole page
