+++ ui/apps/dev-server-ui/src/app/(dashboard)/functions/config/FunctionConfigurationContainer.tsx
@@ -1,8 +1,19 @@
 'use client';
 
+import { useMemo } from 'react';
+import { useRouter } from 'next/navigation';
+import { Button } from '@inngest/components/Button';
 import { FunctionConfiguration } from '@inngest/components/FunctionConfiguration';
+import { Header } from '@inngest/components/Header/Header';
+import { InvokeButton } from '@inngest/components/InvokeButton';
+import { RiCloseLine } from '@remixicon/react';
+import { toast } from 'sonner';
 
-import { useGetFunctionQuery } from '@/store/generated';
+import {
+  FunctionTriggerTypes,
+  useGetFunctionQuery,
+  useInvokeFunctionMutation,
+} from '@/store/generated';
 
 type FunctionDetailsProps = {
   onClose: () => void;
@@ -17,9 +28,51 @@ export function FunctionConfigurationContainer({ onClose, functionSlug }: Functi
     }
   );
 
+  const [invokeFunction] = useInvokeFunctionMutation();
+
+  const inngestFunction = data?.functionBySlug;
+  const triggers = inngestFunction?.triggers;
+
+  const router = useRouter();
+  const doesFunctionAcceptPayload = useMemo(() => {
+    return Boolean(triggers?.some((trigger) => trigger.type === FunctionTriggerTypes.Event));
+  }, [triggers]);
+
   if (isFetching || !data || !data.functionBySlug) {
     return null;
   }
 
-  return <FunctionConfiguration onClose={onClose} inngestFunction={data.functionBySlug} />;
+  const header = (
+    <Header
+      breadcrumb={[{ text: inngestFunction.name }]}
+      action={
+        <div className="flex flex-row items-center justify-end gap-2">
+          <InvokeButton
+            kind="primary"
+            appearance="solid"
+            disabled={false}
+            doesFunctionAcceptPayload={doesFunctionAcceptPayload}
+            btnAction={async ({ data, user }) => {
+              await invokeFunction({
+                data,
+                functionSlug: inngestFunction.slug,
+                user,
+              });
+              toast.success('Function invoked');
+              router.push('/runs');
+            }}
+          />
+          <Button
+            icon={<RiCloseLine className="text-muted h-5 w-5" />}
+            kind="secondary"
+            appearance="ghost"
+            size="small"
+            onClick={() => onClose()}
+          />
+        </div>
+      }
+    />
+  );
+
+  return <FunctionConfiguration inngestFunction={data.functionBySlug} header={header} />;
 }
